pkp/pkp-lib#8409 restructure job and queue classes
