+++ Jobs/Submissions/RemoveSubmissionFromSearchIndexJob.php
@@ -18,10 +18,7 @@ declare(strict_types=1);
 namespace PKP\Jobs\Submissions;
 
 use APP\core\Application;
-use APP\facades\Repo;
-use PKP\Domains\Jobs\Exceptions\JobException;
-
-use PKP\Support\Jobs\BaseJob;
+use PKP\Jobs\BaseJob;
 
 class RemoveSubmissionFromSearchIndexJob extends BaseJob
 {
