+++ classes/job/interfaces/Eloquentable.php
@@ -3,19 +3,18 @@
 declare(strict_types=1);
 
 /**
- * @file Support/Interfaces/Core/Eloquentable.php
+ * @file classes/job/interfaces/Eloquentable.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class Eloquentable
- * @ingroup support
  *
  * @brief Interface for Eloquent Classes
  */
 
-namespace PKP\Support\Interfaces\Core;
+namespace PKP\job\interfaces;
 
 use Illuminate\Contracts\Support\Arrayable;
