+++ classes/job/interfaces/JobRepositoryInterface.php
@@ -3,22 +3,21 @@
 declare(strict_types=1);
 
 /**
- * @file Domains/Jobs/Interfaces/JobRepositoryInterface.php
+ * @file classes/job/interfaces/JobRepositoryInterface.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class JobRepositoryInterface
- * @ingroup support
  *
  * @brief Interface for Domain's Jobs Repository
  */
 
-namespace PKP\Domains\Jobs\Interfaces;
+namespace PKP\job\interfaces;
 
 use Illuminate\Pagination\LengthAwarePaginator;
-use PKP\Support\Interfaces\Core\Repository;
+use PKP\job\interfaces\Repository;
 
 interface JobRepositoryInterface extends Repository
 {
