+++ classes/job/models/Job.php
@@ -3,25 +3,24 @@
 declare(strict_types=1);
 
 /**
- * @file Domains/Jobs/Job.php
+ * @file classes/job/models/Job.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class Job
- * @ingroup domains
  *
  * @brief Laravel Eloquent model for Jobs table
  */
 
-namespace PKP\Domains\Jobs;
+namespace PKP\job\models;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Support\InteractsWithTime;
 use PKP\config\Config;
-use PKP\Domains\Jobs\Traits\Attributes;
-use PKP\Support\Database\Model;
+use PKP\job\models\Model;
+use PKP\job\traits\Attributes;
 
 class Job extends Model
 {
