+++ classes/job/models/Model.php
@@ -3,25 +3,23 @@
 declare(strict_types=1);
 
 /**
- * @file Support/Database/Model.php
+ * @file classes/job/models/Model.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class Model
- * @ingroup support
  *
  * @brief Abstract class for Eloquent Models
  */
 
-namespace PKP\Support\Database;
+namespace PKP\job\models;
 
-use Illuminate\Database\Eloquent\Model as BaseEloquent;
+use Illuminate\Database\Eloquent\Model as BaseModel;
+use PKP\job\interfaces\Eloquentable;
 
-use PKP\Support\Interfaces\Core\Eloquentable;
-
-abstract class Model extends BaseEloquent implements Eloquentable
+abstract class Model extends BaseModel implements Eloquentable
 {
     /**
      * Return model's primary key
