+++ classes/job/repositories/BaseRepository.php
@@ -3,25 +3,23 @@
 declare(strict_types=1);
 
 /**
- * @file Support/Repositories/BaseRepository.php
+ * @file classes/job/repositories/BaseRepository.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class BaseRepository
- * @ingroup support
  *
  * @brief Abstract class BaseRepository
  */
 
-namespace PKP\Support\Repositories;
+namespace PKP\job\repositories;
 
 use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
-
-use PKP\Support\Interfaces\Core\Repository;
+use PKP\job\interfaces\Repository;
 
 abstract class BaseRepository implements Repository
 {
