+++ classes/job/traits/Attributes.php
@@ -3,19 +3,18 @@
 declare(strict_types=1);
 
 /**
- * @file Domains/Jobs/Traits/Attributes.php
+ * @file classes/job/traits/Attributes.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class Attributes
- * @ingroup domains
  *
  * @brief Attributes trait for Jobs model
  */
 
-namespace PKP\Domains\Jobs\Traits;
+namespace PKP\job\traits;
 
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Date;
