+++ classes/queue/WorkerConfiguration.php
@@ -2,19 +2,18 @@
 declare(strict_types=1);
 
 /**
- * @file Domains/Jobs/WorkerConfiguration.php
+ * @file classes/queue/WorkerConfiguration.php
  *
  * Copyright (c) 2014-2022 Simon Fraser University
  * Copyright (c) 2000-2022 John Willinsky
  * Distributed under the GNU GPL v3. For full terms see the file docs/COPYING.
  *
  * @class WorkerConfiguration
- * @ingroup domains
  *
  * @brief Laravel worker configuration manager for managing worker options
  */
 
-namespace PKP\Domains\Jobs;
+namespace PKP\queue;
 
 use Exception;
